@@ -4,10 +4,9 @@ namespace App\Events\Vehiculos;
 
 use App\Models\Notificacion;
 use App\Models\Vehiculos\AsignacionVehiculo;
+use Exception;
 use Illuminate\Broadcasting\Channel;
 use Illuminate\Broadcasting\InteractsWithSockets;
-use Illuminate\Broadcasting\PresenceChannel;
-use Illuminate\Broadcasting\PrivateChannel;
 use Illuminate\Contracts\Broadcasting\ShouldBroadcast;
 use Illuminate\Foundation\Events\Dispatchable;
 use Illuminate\Queue\SerializesModels;
@@ -19,25 +18,26 @@ class NotificarAsignacionVehiculoEvent implements ShouldBroadcast
 
     public AsignacionVehiculo $asignacion;
     public Notificacion $notificacion;
-    public $url = '/asignaciones-vehiculos';
+    public string $url = '/asignaciones-vehiculos';
     public int $destinatario;
+
     /**
      * Create a new event instance.
      *
      * @return void
+     * @throws Exception
      */
     public function __construct($asignacion)
     {
         $this->asignacion = $asignacion;
 
         $this->notificacion = $this->obtenerNotificacion();
-        // Notificacion::crearNotificacion($this->obtenerMensaje(), $this->url, TiposNotificaciones::ASIGNACION_VEHICULO, $asignacion->entrega_id, $asignacion->responsable_id, $asignacion, false);
     }
 
     /**
      * Get the channels the event should broadcast on.
      *
-     * @return \Illuminate\Broadcasting\Channel|array
+     * @return Channel
      */
     public function broadcastOn()
     {
@@ -49,6 +49,9 @@ class NotificarAsignacionVehiculoEvent implements ShouldBroadcast
         return 'asignacion-vehiculo-event';
     }
 
+    /**
+     * @throws Exception
+     */
     public function obtenerNotificacion()
     {
         switch ($this->asignacion->estado) {
