@@ -11,19 +11,22 @@ use App\Models\MaterialEmpleado;
 use App\Models\MaterialEmpleadoTarea;
 use App\Models\TransaccionBodega;
 use Exception;
+use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Log;
 use Illuminate\Validation\ValidationException;
 use Src\Config\EstadosTransacciones;
 use Src\Config\MotivosTransaccionesBodega;
+use Src\Shared\Utils;
+use Throwable;
 
 class ComprobanteController extends Controller
 {
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return JsonResponse
      */
     public function index()
     {
@@ -32,67 +35,54 @@ class ComprobanteController extends Controller
         return response()->json(compact('results'));
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
-    public function store(Request $request)
-    {
-        //
-    }
-
-    /**
-     * Display the specified resource.
-     *
-     * @param  \App\Models\Comprobante  $comprobante
-     * @return \Illuminate\Http\Response
-     */
-    public function show(Comprobante $comprobante)
-    {
-        Log::channel('testing')->info('Log', ['Show de comprobante:', $comprobante]);
-    }
-
 
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\Comprobante  $comprobante
-     * @return \Illuminate\Http\Response
+     * @param ComprobanteRequest $request
+     * @param int $comprobante
+     * @return JsonResponse
+     * @throws Throwable
      */
     public function update(ComprobanteRequest $request, int $comprobante)
     {
-        // Log::channel('testing')->info('Log', ['[Update] El comprobante a modificar es:', $request->all(), $comprobante]);
-        $comprobante = Comprobante::where('transaccion_id', $comprobante)->first();
-        // Log::channel('testing')->info('Log', ['[Despues de update] El comprobante a modificar es:', $comprobante]);
-        if (!$comprobante->firmada) {
-            $datos = $request->validated();
-            $comprobante->update($datos);
-            $comprobante->refresh();
+        try {
+            $comprobante = Comprobante::where('transaccion_id', $comprobante)->first();
+            // Log::channel('testing')->info('Log', ['[Despues de update] El comprobante a modificar es:', $comprobante]);
+            if (!$comprobante->firmada) {
+                DB::beginTransaction();
+                $datos = $request->validated();
+                $comprobante->update($datos);
+                $comprobante->refresh();
 
-            if ($comprobante->firmada) {
-                $transaccion = TransaccionBodega::find($comprobante->transaccion_id);
-                $transaccion->estado_id = EstadosTransacciones::COMPLETA;
-                $transaccion->save();
-                $transaccion->latestNotificacion()->update(['leida' => true]);
-                $esEgresoLiquidacionMateriales = TransaccionBodega::verificarEgresoLiquidacionMateriales($transaccion->motivo_id, $transaccion->motivo->tipo_transaccion_id, MotivosTransaccionesBodega::egresoLiquidacionMateriales);
-                $noGeneraComprobante = TransaccionBodega::verificarMotivosEgreso($transaccion->motivo_id);
-                if (!$esEgresoLiquidacionMateriales && !$noGeneraComprobante) TransaccionBodega::asignarMateriales($transaccion);
-            }
-        } else throw new Exception('Transacción ya fue firmada');
+                if ($comprobante->firmada) {
+                    $transaccion = TransaccionBodega::find($comprobante->transaccion_id);
+                    $transaccion->estado_id = EstadosTransacciones::COMPLETA;
+                    $transaccion->save();
+                    $transaccion->latestNotificacion()->update(['leida' => true]);
+                    $es_egreso_liquidacion_materiales = TransaccionBodega::verificarEgresoLiquidacionMateriales($transaccion->motivo_id, $transaccion->motivo->tipo_transaccion_id, MotivosTransaccionesBodega::egresoLiquidacionMateriales);
+                    $no_genera_comprobante = TransaccionBodega::verificarMotivosEgreso($transaccion->motivo_id);
+                    if (!$es_egreso_liquidacion_materiales && !$no_genera_comprobante) TransaccionBodega::asignarMateriales($transaccion);
+                }
+                DB::commit();
+            } else throw new Exception('Transacción ya fue firmada');
 
-        $modelo = new ComprobanteResource($comprobante);
-        $mensaje = 'Comprobante actualizado correctamente';
-        return response()->json(compact('mensaje', 'modelo'));
+            $modelo = new ComprobanteResource($comprobante);
+            $mensaje = 'Comprobante actualizado correctamente';
+            return response()->json(compact('mensaje', 'modelo'));
+        } catch (Throwable $th) {
+            DB::rollBack();
+            throw Utils::obtenerMensajeErrorLanzable($th, 'No se pudo completar la operación');
+        }
     }
+
     /**
      * Actualizar parcial la recepcion del egreso.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\TransaccionBodega  $transaccion
-     * @return \Illuminate\Http\Response
+     * @param Request $request
+     * @param TransaccionBodega $transaccion
+     * @return JsonResponse
+     * @throws Throwable
      */
     public function comprobanteParcial(Request $request, TransaccionBodega $transaccion)
     {
@@ -105,12 +95,12 @@ class ComprobanteController extends Controller
                 if ($detalle) {
                     $detalle->recibido += $item['recibido'];
                     $detalle->save();
-                    $itemInventario = Inventario::find($item['id']);
-                    if ($itemInventario) {
+                    $item_inventario = Inventario::find($item['id']);
+                    if ($item_inventario) {
                         if ($transaccion->tarea_id) {
-                            MaterialEmpleadoTarea::cargarMaterialEmpleadoTarea($itemInventario->detalle_id, $transaccion->responsable_id, $transaccion->tarea_id, $item['recibido'], $transaccion->cliente_id, $transaccion->proyecto_id, $transaccion->etapa_id);
+                            MaterialEmpleadoTarea::cargarMaterialEmpleadoTarea($item_inventario->detalle_id, $transaccion->responsable_id, $transaccion->tarea_id, $item['recibido'], $transaccion->cliente_id, $transaccion->proyecto_id, $transaccion->etapa_id);
                         } else {
-                            MaterialEmpleado::cargarMaterialEmpleado($itemInventario->detalle_id, $transaccion->responsable_id, $item['recibido'], $transaccion->cliente_id);
+                            MaterialEmpleado::cargarMaterialEmpleado($item_inventario->detalle_id, $transaccion->responsable_id, $item['recibido'], $transaccion->cliente_id);
                         }
                     } else {
                         throw new Exception('No se encontró el item de inventario');
@@ -120,7 +110,7 @@ class ComprobanteController extends Controller
                 }
             }
 
-            if (Comprobante::verificarEgresoCompletado($transaccion->id)) { //se completo el egreso 
+            if (Comprobante::verificarEgresoCompletado($transaccion->id)) { //se completó el egreso
                 $transaccion->estado_id = EstadosTransacciones::COMPLETA;
                 $transaccion->observacion_est = $request->observacion;
                 $transaccion->save();
@@ -149,21 +139,9 @@ class ComprobanteController extends Controller
             Log::channel('testing')->info('Log', ['error:', $e]);
             DB::rollback();
             throw ValidationException::withMessages(['error' => [$e->getMessage() . '. ' . $e->getLine()]]);
-            return response()->json(['mensaje' => 'Ha ocurrido un error al insertar el registro' . $e->getMessage() . $e->getLine()], 422);
-            return response()->json(['mensaje' => $e], 422);
         }
 
         return response()->json(compact('mensaje', 'modelo'));
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  \App\Models\Comprobante  $comprobante
-     * @return \Illuminate\Http\Response
-     */
-    public function destroy(Comprobante $comprobante)
-    {
-        //
-    }
 }
